Don’t try and match event id on mouse up
Stop any ongoing moves.  This is to fix an intermitent issue on Bukwild.com where you would get stuck in a state where pointerActive was true but the current touch id didn’t match the saved one.  I was never able to determine how to get into that state, but it was relatively common with enough interactions.